requirements.txt
pytube
openai==0.28
transformers
torch
youtube-transcript-api 
python-dotenv 
